+++ lesson_04/shapeMain.go
@@ -1,15 +1,17 @@
 package main
 
 import (
-	"lesson_04/shapepkg"
+	"lesson_04/shapedescription"
 )
 
 func main() {
-	c := shapepkg.Circle{Radius: 8.0}
-	r := shapepkg.Rectangle{
+	c := shapedescription.Circle{Radius: 8.0}
+	r := shapedescription.Rectangle{
 		Height: 9.0,
 		Width:  3.0,
 	}
-	shapepkg.DescribeShape(c)
-	shapepkg.DescribeShape(r)
+	q := shapedescription.Square{Length: 2.0}
+	shapedescription.DescribeShape(c)
+	shapedescription.DescribeShape(r)
+	shapedescription.DescribeShape(q)
 }
